// torus-sq-indexer/src/types/models/Block.ts
export type BlockProps = Omit<Block, NonNullable<FunctionPropertyNames<Block>> | '_name'>;

/*
 * Compat types allows for support of alternative `id` types without refactoring the node
 */
type CompatBlockProps = Omit<BlockProps, 'id'> & { id: string; };
type CompatEntity = Omit<Entity, 'id'> & { id: string; };

export class Block implements CompatEntity {

    constructor(
        
        id: string,
        height: number,
        eventCount: number,
        extrinsicCount: number,
        hash: string,
        parentHash: string,
        specVersion: number,
        timestamp: Date,
    ) {
        this.id = id;
        this.height = height;
        this.eventCount = eventCount;
        this.extrinsicCount = extrinsicCount;
        this.hash = hash;
        this.parentHash = parentHash;
        this.specVersion = specVersion;
        this.timestamp = timestamp;
        
    }

    public id: string;
    public height: number;
    public eventCount: number;
    public extrinsicCount: number;
    public hash: string;
    public parentHash: string;
    public specVersion: number;
    public timestamp: Date;
    

    get _name(): string {
        return 'Block';
    }

    async save(): Promise<void> {
        const id = this.id;
        assert(id !== null, "Cannot save Block entity without an ID");
        await store.set('Block', id.toString(), this as unknown as CompatBlockProps);
    }

    static async remove(id: string): Promise<void> {
        assert(id !== null, "Cannot remove Block entity without an ID");
        await store.remove('Block', id.toString());
    }

    static async get(id: string): Promise<Block | undefined> {
        assert((id !== null && id !== undefined), "Cannot get Block entity without an ID");
        const record = await store.get('Block', id.toString());
        if (record) {
            return this.create(record as unknown as BlockProps);
        } else {
            return;
        }
    }

    static async getByHeight(height: number, options: GetOptions<CompatBlockProps>): Promise<Block[]> {
        // Inputs must be cast as the store interface has not been updated to support alternative ID types
        const records = await store.getByField<CompatBlockProps>('Block', 'height', height, options);
        return records.map(record => this.create(record as unknown as BlockProps));
    }
    

    static async getByHash(hash: string, options: GetOptions<CompatBlockProps>): Promise<Block[]> {
        // Inputs must be cast as the store interface has not been updated to support alternative ID types
        const records = await store.getByField<CompatBlockProps>('Block', 'hash', hash, options);
        return records.map(record => this.create(record as unknown as BlockProps));
    }
    

    static async getByTimestamp(timestamp: Date, options: GetOptions<CompatBlockProps>): Promise<Block[]> {
        // Inputs must be cast as the store interface has not been updated to support alternative ID types
        const records = await store.getByField<CompatBlockProps>('Block', 'timestamp', timestamp, options);
        return records.map(record => this.create(record as unknown as BlockProps));
    }
    


    /**
     * Gets entities matching the specified filters and options.
     *
     * ⚠️ This function will first search cache data followed by DB data. Please consider this when using order and offset options.⚠️
     * */
    static async getByFields(filter: FieldsExpression<BlockProps>[], options: GetOptions<BlockProps>): Promise<Block[]> {
        const records = await store.getByFields<CompatBlockProps>('Block', filter  as unknown as FieldsExpression<CompatBlockProps>[], options as unknown as GetOptions<CompatBlockProps>);
        return records.map(record => this.create(record as unknown as BlockProps));
    }

    static create(record: BlockProps): Block {
        assert(record.id !== undefined && record.id !== null, "id must be provided");
        const entity = new this(
            record.id,
            record.height,
            record.eventCount,
            record.extrinsicCount,
            record.hash,
            record.parentHash,
            record.specVersion,
            record.timestamp,
        );
        Object.assign(entity,record);
        return entity;
    }
}